+++ src/edu/umn/cs/spatial/mapReduce/KNNPointRecordReader.java
@@ -17,7 +17,7 @@ import edu.umn.edu.spatial.Point;
  * @author aseldawy
  *
  */
-public class KNNPointRecordReader implements RecordReader<IntWritable, Point> {
+public class KNNPointRecordReader implements RecordReader<LongWritable, Point> {
   
   private RecordReader<LongWritable, Text> lineRecordReader;
   private LongWritable subKey;
@@ -35,7 +35,7 @@ public class KNNPointRecordReader implements RecordReader<IntWritable, Point> {
 	 * It consumes the end of line also when reading the rectangle.
 	 * It stops after reading the first end of line (after) end.
 	 */
-	public boolean next(IntWritable key, Point value) throws IOException {
+	public boolean next(LongWritable key, Point value) throws IOException {
 	  if (!lineRecordReader.next(subKey, subValue) || subValue.getLength() < 4) {
 	    // Stop on wrapped reader EOF or a very short line which indicates EOF too
 	    return false;
@@ -43,7 +43,7 @@ public class KNNPointRecordReader implements RecordReader<IntWritable, Point> {
 	  // Convert to a regular string to be able to use split
 	  String line = new String(subValue.getBytes(), 0, subValue.getLength());
 	  String[] parts = line.split(",");
-	  key.set(Integer.parseInt(parts[0]));
+	  key.set(Long.parseLong(parts[0]));
 	  value.id = key.get();
 	  value.x = Integer.parseInt(parts[1]);
 	  value.y = Integer.parseInt(parts[2]);
@@ -65,9 +65,9 @@ public class KNNPointRecordReader implements RecordReader<IntWritable, Point> {
   }
 
   @Override
-  public IntWritable createKey() {
+  public LongWritable createKey() {
     subKey = lineRecordReader.createKey();
-    return new IntWritable();
+    return new LongWritable();
   }
 
   @Override
